First iteration at moving ProfileActivity logic into fragments
