+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/ui/activity/AnotherUserProfileActivity.java
@@ -12,6 +12,7 @@ import android.support.v7.app.ActionBar;
 import com.microsoft.embeddedsocial.account.UserAccount;
 import com.microsoft.embeddedsocial.ui.activity.base.BaseProfileActivity;
 import com.microsoft.embeddedsocial.service.IntentExtras;
+import com.microsoft.embeddedsocial.ui.fragment.AnotherUserProfileFragment;
 
 /**
  * Shows another user's profile.
@@ -19,6 +20,12 @@ import com.microsoft.embeddedsocial.service.IntentExtras;
 public class AnotherUserProfileActivity extends BaseProfileActivity {
 
 	private String userHandle;
+
+	@Override
+	protected void setupFragments() {
+		setActivityContent(new AnotherUserProfileFragment());
+		super.setupFragments();
+	}
 
 	@Override
 	protected void onPostCreate(Bundle savedInstanceState) {
@@ -31,15 +38,6 @@ public class AnotherUserProfileActivity extends BaseProfileActivity {
 		super.setupActionBar(actionBar);
 		String name = getIntent().getStringExtra(IntentExtras.NAME);
 		actionBar.setTitle(name);
-	}
-
-	@Override
-	protected void initExtraVariables() {
-		Intent intent = getIntent();
-		setUserHandle(intent.getStringExtra(IntentExtras.USER_HANDLE));
-		setUserName(intent.getStringExtra(IntentExtras.NAME));
-		setIsCurrentUser(false);
-		setFeedIsReadable(!intent.getBooleanExtra(IntentExtras.FEED_IS_NOT_READABLE, false));
 	}
 
 	@Override
